requirements.txt
numpy>=1.24.0
pillow>=10.0.0
onnxruntime>=1.16.0
rich>=13.0.0
pyfiglet>=0.8.0
